--- src/br/edu/ifal/commanage/commanage-master/src/br/edu/ifal/commanage/bo/CompanyBO.java
@@ -1,46 +1,0 @@
-package br.edu.ifal.commanage.bo;
-
-import br.edu.ifal.commanage.model.Company;
-import br.edu.ifal.commanage.util.FieldValidation;
-
-public class CompanyBO {
-	
-	private Company company;
-	
-	public CompanyBO(Company company) {
-		this.company = company;
-	}
-	
-	public String getName() {
-		return company.getName();
-	}
-	
-	public void validateSetCompanyName(String name) throws Exception {
-		if (FieldValidation.isValidName(name))
-			company.setName(name);
-		else
-			throw new Exception("Invalid name");
-	}
-	
-	public String getCnpj() {
-		return company.getCnpj();
-	}
-	
-	public void validateSetCompanyCnpj(String cnpj) throws Exception {
-		if (FieldValidation.isValidCnpj(cnpj))
-			company.setCnpj(cnpj);
-		else
-			throw new Exception("Invalid cnpj");
-	}
-	
-	public String getPhone() {
-		return company.getPhone();
-	}
-	
-	public void validateSetCompanyPhone(String phone) throws Exception {
-		if (FieldValidation.isValidPhone(phone))
-			company.setPhone(phone);
-		else
-			throw new Exception("Invalid phone");
-	}
-}
