--- src/br/edu/ifal/commanage/commanage-master/src/br/edu/ifal/commanage/model/Company.java
@@ -1,32 +1,0 @@
-package br.edu.ifal.commanage.model;
-
-public class Company {
-	
-	private String name;
-	private String cnpj;
-	private String phone;
-	
-	public String getName() {
-		return name;
-	}
-	
-	public void setName(String name) {
-		this.name = name;
-	}
-	
-	public String getCnpj() {
-		return cnpj;
-	}
-	
-	public void setCnpj(String cnpj) {
-		this.cnpj = cnpj;
-	}
-	
-	public String getPhone() {
-		return phone;
-	}
-	
-	public void setPhone(String phone) {
-		this.phone = phone;
-	}
-}
