--- src/br/edu/ifal/commanage/commanage-master/src/br/edu/ifal/commanage/model/Worker.java
@@ -1,34 +1,0 @@
-package br.edu.ifal.commanage.model;
-
-public class Worker {
-	
-	private String name;
-	private String phone;
-	private String email;
-	
-	public String getName() {
-		return name;	
-	}
-	
-	public void setName(String name) {
-		this.name = name;
-	}
-	
-	public String getPhone() {
-		return phone;
-	}
-	
-	public void setPhone(String phone) {
-		this.phone = phone;
-	}
-	
-	public String getEmail() {
-		return email;
-	}
-	
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	
-	
-}
